feat(deployer): replace delete_service by stop_service
According to [this
comment](https://github.com/shuttle-hq/shuttle/pull/616#issuecomment-1421255649),
it's not necessary to remove the service from the database and we only
need to stop it. So we can rename the `delete_service` function to
`stop_service` and just remove the persistence line for dropping the
service in database.

Signed-off-by: Jérémie Drouet <jeremie.drouet@gmail.com>