review-feedback: adjust shuttle admin args...
...and removed Makefile containters additional build args for
panamax and postgres. It seems there is a 'PLATFORMS' env variable
in the Makefile which can be set before building the images and used
by `buildx` to build against the mentioned target platforms.

The `PLATFORMS` env variable is not needed on my machine because I use
Docker Desktop, which supports running containers built for other arch
using behind the scenes a LinuxKit VM that intercepts [1][2] the binaries
and runs them using binfmt_misc, through QEMU [1].

[1] https://www.docker.com/blog/the-magic-behind-the-scenes-of-docker-desktop/
[2] https://stackoverflow.com/questions/72444103/what-does-running-the-multiarch-qemu-user-static-does-before-building-a-containe

Signed-off-by: Iulian Barbu <iulianbarbu2@gmail.com>