feat(gateway, cargo-shuttle): implement pagination for `project list`
This change adds query parameters to the `/projects` endpoint that allow
specifying an `offset` and a `limit` to allow for pagination. The
`cargo-shuttle` CLI has also been updated to take those as optional
parameters for the `project list` subcommand with default values of `0`
for `offset` and `10` for `limit`.

refactor(deployer, cargo-shuttle): implement pagination for `deployment list`

This adds pagination functionality to `cargo shuttle deployment list`
just like `cargo shuttle project list` and allows the same parameters.