Add commands cookbook to cargo-shuttle --help
This commit adds a couple of example `cookbook` style commands to the
`--help` subcommand of `cargo-shuttle`. It took some time for me to
realise the difference between `deploy` and `deployments`, or to
discover that I could watch my logs in real time via `logs --follow`. My
hope is that this extended help assists newcomers in figuring out what
cargo-shuttle can do.