common: renamed ProjectName to RawProjectName
The common::ProjectName clashed with the gateway::ProjectName and looked similar,
so I think it would be best to clarify the difference between them through naming.
If this change will not be merged then the clashing will still occur for the ProjectName
introduced in the orchestrator.

I chose `Raw` prefix because it seems that this struct is used by multiple crates
and the other can build on top of it.