deployer: runtime almost started
The runtime doesn't passes the started state because it can not pass the
health check. I tested with a shuttle-runtime that doesn't have the implementation
for the health check RPC.

TODO: build a shuttle-runtime based on the runtime from this branch and try to
deploy it againt on the deployer.