feat(runtime): add alpha runtime version check (#1088)
* feat(runtime): add alpha runtime version check

Add a version flag check to the alpha runtime and
check if its version is the same as the version
of cargo-shuttle. Warn the user if that is not the
case and give hints. The next runtime already has
such a check.

Closes #1053

* fix: apply lint from cargo clippy

* fix: check version compatibility and improve hints

Use cargo's rule for checking if two versions are
compatible. Give hints on how to update the older
componenent.

* fix: share `print_version` between both runtimes

Use the same function in 'alpha' and 'next' for
printing the current version of shuttle-runtime.

* fix: make `print_version` pub

Allow use of `print_version` in next binary by
making `print_version` available to it, too.

* fix: remove duplication and apply lints

* Update cargo-shuttle/src/lib.rs

Co-authored-by: Oddbjørn Grødem <29732646+oddgrd@users.noreply.github.com>

---------

Co-authored-by: Iulian Barbu <14218860+iulianbarbu@users.noreply.github.com>
Co-authored-by: Oddbjørn Grødem <29732646+oddgrd@users.noreply.github.com>