fix(common): sqlx compilation issue
When the entire workspace is compiled together,
like in CI clippy check, logger enables the sqlx/postgres
feature. When ProjectName in common is compiled, it is usually
done with only the sqlx/sqlite feature. Enabling postgres from
logger makes the sqlx derive macro break for some reason.
This fixes it.