Improve design of `generate_project`
The different parts of generating the template
and their use of the arguments passed to
`generate_project` are now divided more clearly.

This decreases the size and overhead of the implementation.

The tests checking the paring are also gone. Instead, the
cloning and the copying of templates is tested. The
integration tests should be sufficient for ensuring that
the paring in this implementation remains correct.