bug: tests losing track of account tier
While implementing the ability to allow pro users to start project above
the soft limit I could not understand why the tests kept failing.
Eventually found is was because gateway thought these were basic users.
This happened because our mock auth only tracked scopes and not the
actual tiers of users it managed.